Implement update storing/lookup and function state handling
A new FunctionState class hides away all the details of storing location-value
pairs. The implementation is relatively simple atm but will be extended in future
to allow observation, ...

A Storage class uses function states and handles the lookup and manipulation of
states. Internally the program functions and other ASM functions are splitted into
two function states, because the scheduler needs an efficient way to iterate
through all existing program locations (e.g. when determining the set of runnable
agents).