Use explicit merge in update set fork guard
RAII merge was nice but has the big disadvantage, that the deconstructors may
throw a merge conflict exception. This is especially bad when another exception
is already propagated and a merge conflict occurs. In this case the application
will terminate immediately.

Using an explicit merge avoids this problem. Later the explicit merge can be used
for automatic rollback in case of CASM exception handling.