Better error messages on merge conflicts esp. when using the import feature
When using the import feature it could happen that conflicting locations have
different function identifiers which all refer to the same function definition.
This happens because of the namespacing provided by import. In such cases the
error message producer cannot assume that the location string of the existing and
conflicting update is the same.

The error message producer has been refactored and now generates the location
string of each update separately.

Additionally some code has been shared which was the same in all error message
producers.