Remove superfluous reference atom checks from ConsistencyCheckPass
The symbol resolver assures that all reference atoms are resolved (if this is
not possible, the symbol resolver will throw an error), thus no symbol table
checking is required in ConsistencyCheckPass.

If something went wrong in the previous passes the ConsistencyCheckPass will
still detect it because the type will be missing.