Always fork an update set when requested
Without detailed analysis it is dangerous to perform this kind of optimization.
E.g. it will fail as soon as we implement exception handling:
```
{| // S1
   x := 2
   try {| // S2
      x := 3
      throw ...
   |} catch ...
   assert( x = 2 )
|}
```
The optimization will no fork the inner seq. update set S2 an thus apply the
update "x := 3" to the update set S1. The assertion `x=2` will fail.

This optimization should be performed in the IR instead.