TypeCheckPass: Drop all the typing related code
Typing should only be done in the TypeInferencePass. This removes a huge amount
of duplicated code and also makes it possible to implement the `Type Definition`
feature properly. So that it also works for enums and so on.

The relative path resolving will also be added to the type inference, so that
relative enum contants can be used anywhere and not only where it has been
hardcoded (case, let).