[GrammarParser] Use SimpleOrClaspedTerm instead of Term for TypeCastingExpression
Using the example `1 + 1 as Boolean`:

Casting of 1 + 1 is ambiguous because it can either be `1 + (1 as Boolean)`
or `(1 + 1) as Boolean` depending on the precendence of the operator.

Better avoid this and force brackets if the second one was intended, by using
the SimpleOrClaspedTerm.

E.g. `assert( c = 0b1111'0000 as Binary'16 )` is such an real-world example
      where the result of = was mistakenly casted to Binary'16.

This change also made is possible to drop the precendence enforcement of the
`as` term.