[GrammarParser, GrammarLexer] Fix incorrect parsing of relative paths
Example:
case c of {
    ::Red: bar
    ::Blue: skip
}

The problem was, that the parser treated the bar rule call and the immediately
following relative path ::Blue as a single absolute path bar::Blue.

This happened because the grammar allowed to write identifier paths with spaces
between the segments, e.g. "Color :: Red". By removing the ability to add spaces
between the segments the example above is correctly parsed.

Fixes casm-lang/casm#59