TypeInferencePass: Drop all the "no type found" errors
This hides some errors when the late symbol resolving fails (checked in the
consistency checker). Also the error message is on the wrong level, because
the real error is somewhere in the expression. Thus duplicated error messages
may be shown - one about the real error and one useless "no type found".

If there is a bug in the type inference, then the consistency checker will also
catch it, because the consistency checker verifies that all expressions are
typed. No need to duplicate the checks in the type inference.