TypeInferencePass: Drop all the "no type found" errors from ForallRule
This only prints duplicated error messages - one about the real error and one
useless "no type found". If there is a bug in the type inference, then the
consistency checker will also catch it, because the consistency checker verifies
that all expressions are typed. No need to duplicate the checks in the type
inference.