[import] Add a loading strategy interface + a file loading strategy implementation
The loading strategy can be used to load a source code for a given identifier
path. The resolving of the identifier path depends on the actual implementation.
For example the file loading strategy will convert it into a file-system path
by replacing all dots with slashed + adding the base path.